+++ tests/timeseries/test_ringbuffer_serialization.py
@@ -12,15 +12,16 @@ from typing import Any
 import numpy as np
 import pytest
 
+import frequenz.sdk.timeseries._ringbuffer_serialization as io
 from frequenz.sdk.timeseries import Sample
-from frequenz.sdk.timeseries._serializable_ringbuffer import SerializableRingBuffer
+from frequenz.sdk.timeseries._ringbuffer import OrderedRingBuffer
 
 FIVE_MINUTES = timedelta(minutes=5)
 _29_DAYS = 60 * 24 * 29
 ONE_MINUTE = timedelta(minutes=1)
 
 
-def load_dump_test(dumped: SerializableRingBuffer[Any]) -> None:
+def load_dump_test(dumped: OrderedRingBuffer[Any], path: str) -> None:
     """Test ordered ring buffer."""
     size = dumped.maxlen
 
@@ -42,11 +43,11 @@ def load_dump_test(dumped: SerializableRingBuffer[Any]) -> None:
             )
         )
 
-    dumped.dump()
+    io.dump(dumped, path)
 
     # Load old data
     # pylint: disable=protected-access
-    loaded = SerializableRingBuffer.load(dumped._path)
+    loaded = io.load(path)
     assert loaded is not None
 
     np.testing.assert_equal(dumped[:], loaded[:])
@@ -72,21 +73,21 @@ def test_load_dump_short(tmp_path_factory: pytest.TempPathFactory) -> None:
     tmpdir = tmp_path_factory.mktemp("load_dump")
 
     load_dump_test(
-        SerializableRingBuffer(
+        OrderedRingBuffer(
             [0.0] * int(24 * FIVE_MINUTES.total_seconds()),
             FIVE_MINUTES,
-            f"{tmpdir}/test_list.bin",
             datetime(2, 2, 2, tzinfo=timezone.utc),
-        )
+        ),
+        f"{tmpdir}/test_list.bin",
     )
 
     load_dump_test(
-        SerializableRingBuffer(
+        OrderedRingBuffer(
             np.empty(shape=(24 * int(FIVE_MINUTES.total_seconds()),), dtype=np.float64),
             FIVE_MINUTES,
-            f"{tmpdir}/test_array.bin",
             datetime(2, 2, 2, tzinfo=timezone.utc),
-        )
+        ),
+        f"{tmpdir}/test_array.bin",
     )
 
 
@@ -95,19 +96,19 @@ def test_load_dump(tmp_path_factory: pytest.TempPathFactory) -> None:
     tmpdir = tmp_path_factory.mktemp("load_dump")
 
     load_dump_test(
-        SerializableRingBuffer(
+        OrderedRingBuffer(
             [0.0] * _29_DAYS,
             ONE_MINUTE,
-            f"{tmpdir}/test_list_29.bin",
             datetime(2, 2, 2, tzinfo=timezone.utc),
-        )
+        ),
+        f"{tmpdir}/test_list_29.bin",
     )
 
     load_dump_test(
-        SerializableRingBuffer(
+        OrderedRingBuffer(
             np.empty(shape=(_29_DAYS,), dtype=np.float64),
             ONE_MINUTE,
-            f"{tmpdir}/test_array_29.bin",
             datetime(2, 2, 2, tzinfo=timezone.utc),
-        )
+        ),
+        f"{tmpdir}/test_array_29.bin",
     )