+++ RELEASE_NOTES.md
@@ -17,6 +17,7 @@
 * Add a resampler in the MovingWindow to control the granularity of the samples to be stored in the underlying buffer (#269).
   Notice that the parameter `sampling_period` has been renamed to `input_sampling_period`
   to better distinguish it from the sampling period parameter in the resampler.
+* The serialization feature for the ringbuffer was made more flexible. The `dump` and `load` methods can now work directly with a ringbuffer instance.
 
 ## New Features